working on remaking UserPreferences page
