+++ src/css/UserDashboardPage.css
@@ -1,6 +1,7 @@
 .account-wrapper {
   display: flex;
-  justify-content: space-between;
+  /* justify-content: space-between; */
+  min-height: calc(100vh - var(--nav-height));
 }
 
 .account-page {
@@ -12,7 +13,7 @@
   display: flex;
   justify-content: center;
   background: var(--offwhite);
-  padding: 0 25px;  
+  padding: 0 25px;
   width: 15%;
 }
 
@@ -27,7 +28,7 @@
 .account-nav ul li a {
   text-decoration: none;
   font-size: 18px;
-  color: var(--darkgrey)
+  color: var(--darkgrey);
 }
 
 .switch {
@@ -37,17 +38,17 @@
   height: 34px;
 }
 
-.switch input[type=checkbox] {
+.switch input[type='checkbox'] {
   opacity: 0;
   width: 0;
   height: 0;
 }
 
-.switch input[type=checkbox]:checked + .slider {
+.switch input[type='checkbox']:checked + .slider {
   background: green;
 }
 
-.switch input[type=checkbox]:checked + .slider:before {
+.switch input[type='checkbox']:checked + .slider:before {
   -webkit-transform: translateX(26px);
   -ms-transform: translateX(26px);
   transform: translateX(26px);
@@ -63,20 +64,20 @@
   right: 0;
   bottom: 0;
   background-color: var(--red);
-  -webkit-transition: .4s;
-  transition: .4s;
+  -webkit-transition: 0.4s;
+  transition: 0.4s;
 }
 
 .slider:before {
   position: absolute;
-  content: "";
+  content: '';
   left: 4px;
   bottom: 4px;
   width: 26px;
   height: 26px;
   background-color: white;
-  transition: .4s;
-  -webkit-transition: .4s;
+  transition: 0.4s;
+  -webkit-transition: 0.4s;
   border-radius: 50%;
 }
 
@@ -100,7 +101,7 @@
 
 .single-genre {
   font-size: 17px;
-  color: var(--offwhite)
+  color: var(--offwhite);
 }
 
 /*
@@ -116,7 +117,6 @@
   color: white;
   padding-right: 5px;
 }
-
 
 /*
 ======================= Toggle Switch ===========================
@@ -141,7 +141,6 @@
   color: white;
   font-size: 20px;
 }
-
 
 /*
 ======================= Genre List ===========================
@@ -188,12 +187,12 @@
 
 /* When the checkbox is checked, add a blue background */
 .checkmark-container input:checked ~ .checkmark {
-  background-color: #2196F3;
+  background-color: #2196f3;
 }
 
 /* Create the checkmark/indicator (hidden when not checked) */
 .checkmark:after {
-  content: "";
+  content: '';
   position: absolute;
   display: none;
 }
